Refactor taskinstances filter to a separate component.
